packages/core/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.24.0
redis>=5.0.0
requests>=2.31.0
pydantic>=2.0.0
cryptography>=41.0.0
httpx>=0.27.0
google-generativeai>=0.3.0
pyyaml>=6.0
